research-api/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
wikipedia
requests
